+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * New function `function_arguments()` that returns the names of function arguments.
 
+* New function `stationary_distribution()` that computes the stationary distribution corresponding to a transition probability matrix.
+
 # oeli 0.1.0
 
 * Initial CRAN submission.